Revert "Revert "Added interfaces for DAO classes""
This reverts commit 3b4bc8a99146d27bcffde74e0aad1c53111c3bd3.
